+++ SME_Implementations/SingleCycleRISCV/PC_Buses.cs
@@ -5,13 +5,13 @@ namespace SingleCycleRISCV {
     [TopLevelInputBus]
     public interface PC_Input : IBus {
         [InitialValue]
-        uint Address {get; set;}
+        ulong Address {get; set;}
     }
 
     [TopLevelOutputBus]
     public interface ProgramCounter_To_InstructionMemory : IBus {
         [InitialValue]
-        uint Address {get; set;}
+        ulong Address {get; set;}
     }
 
 }
